Reorganizing Git Repo
I tried opening the repo on a different computer and it broke the repository when I tried to make a change...
