+++ .gitignore
@@ -1,4 +1,4 @@
 
-.vscode/
+.vs/
 
-x64/
+*.sln
