--- .vs/VSWorkspaceState.json
@@ -1,7 +1,0 @@
-{
-  "ExpandedNodes": [
-    ""
-  ],
-  "SelectedNode": "\\NeuralNetwork.h",
-  "PreviewInSolutionExplorer": false
-}
