+++ NumberRecognitionAI/.gitignore
@@ -1,0 +1,4 @@
+
+.vscode/
+
+x64/
