+++ NumberRecognitionAI/README.md
@@ -1,0 +1,2 @@
+# NumberRecognitionAI
+A simple educational project for making an AI ML system using only C++ for learning purposes.
