--- README.md
@@ -1,2 +1,0 @@
-# NumberRecognitionAI
-A simple educational project for making an AI ML system using only C++ for learning purposes.
